# docs/swagger.yaml
definitions:
...
  controller.GroupResponse:
    properties:
      accessed_at:
        type: string
      created_at:
        type: string
      id:
        type: string
      request_count:
        type: integer
      rpm:
        additionalProperties:
          type: integer
        type: object
      rpm_ratio:
        type: number
      status:
        type: integer
      tpm:
        additionalProperties:
          type: integer
        type: object
      tpm_ratio:
        type: number
      used_amount:
        type: number
    type: object
...
  middleware.APIResponse:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
...
  model.Group:
    properties:
      created_at:
        type: string
      id:
        type: string
      request_count:
        type: integer
      rpm:
        additionalProperties:
          type: integer
        type: object
      rpm_ratio:
        type: number
      status:
        type: integer
      tpm:
        additionalProperties:
          type: integer
        type: object
      tpm_ratio:
        type: number
      used_amount:
        type: number
    type: object
...
info:
  contact: {}
paths:
...
  /api/group/{group}:
    delete:
      description: Deletes a group by its name
      parameters:
      - description: Group name
        in: path
        name: group
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/middleware.APIResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete a group
      tags:
      - group
    get:
      description: Returns detailed information about a specific group
      parameters:
      - description: Group name
        in: path
        name: group
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/middleware.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/controller.GroupResponse'
              type: object
      security:
      - ApiKeyAuth: []
      summary: Get a group
      tags:
      - group
    post:
      consumes:
      - application/json
      description: Creates a new group with the given information
      parameters:
      - description: Group name
        in: path
        name: group
        required: true
        type: string
      - description: Group information
        in: body
        name: data
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/middleware.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.Group'
              type: object
      security:
      - ApiKeyAuth: []
      summary: Create a new group
      tags:
      - group
    put:
      consumes:
      - application/json
      description: Updates an existing group with the given information
      parameters:
      - description: Group name
        in: path
        name: group
        required: true
        type: string
      - description: Updated group information
        in: body
        name: data
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/middleware.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.Group'
              type: object
      security:
      - ApiKeyAuth: []
      summary: Update a group
      tags:
      - group
  /api/group/{group}/rpm:
    post:
      consumes:
      - application/json
      description: Updates the RPM (Requests Per Minute) for a group
      parameters:
      - description: Group name
        in: path
        name: group
        required: true
        type: string
      - description: RPM information
        in: body
        name: data
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/middleware.APIResponse'
      security:
      - ApiKeyAuth: []
      summary: Update group RPM
      tags:
      - group
  /api/group/{group}/rpm_ratio:
    post:
      consumes:
      - application/json
      description: Updates the RPM (Requests Per Minute) ratio for a group
      parameters:
      - description: Group name
        in: path
        name: group
        required: true
        type: string
      - description: RPM ratio information
        in: body
        name: data
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/middleware.APIResponse'
      security:
      - ApiKeyAuth: []
      summary: Update group RPM ratio
      tags:
      - group
  /api/group/{group}/status:
    post:
      consumes:
      - application/json
      description: Updates the status of a group
      parameters:
      - description: Group name
        in: path
        name: group
        required: true
        type: string
      - description: Status information
        in: body
        name: status
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/middleware.APIResponse'
      security:
      - ApiKeyAuth: []
      summary: Update group status
      tags:
      - group
  /api/group/{group}/tpm:
    post:
      consumes:
      - application/json
      description: Updates the TPM (Tokens Per Minute) for a group
      parameters:
      - description: Group name
        in: path
        name: group
        required: true
        type: string
      - description: TPM information
        in: body
        name: data
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/middleware.APIResponse'
      security:
      - ApiKeyAuth: []
      summary: Update group TPM
      tags:
      - group
  /api/group/{group}/tpm_ratio:
    post:
      consumes:
      - application/json
      description: Updates the TPM (Tokens Per Minute) ratio for a group
      parameters:
      - description: Group name
        in: path
        name: group
        required: true
        type: string
      - description: TPM ratio information
        in: body
        name: data
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/middleware.APIResponse'
      security:
      - ApiKeyAuth: []
      summary: Update group TPM ratio
      tags:
      - group
  /api/groups:
    get:
      description: Returns a list of all groups with pagination
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/middleware.APIResponse'
            - properties:
                data:
                  additionalProperties:
                    allOf:
                    - {}
                    - properties:
                        groups:
                          items:
                            $ref: '#/definitions/controller.GroupResponse'
                          type: array
                        total:
                          type: integer
                      type: object
                  type: object
              type: object
      security:
      - ApiKeyAuth: []
      summary: Get all groups
      tags:
      - groups
  /api/groups/batch_delete:
    post:
      consumes:
      - application/json
      description: Deletes multiple groups by their IDs
      parameters:
      - description: Group IDs
        in: body
        name: ids
        required: true
        schema:
          items:
            type: string
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/middleware.APIResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete multiple groups
      tags:
      - groups
  /api/groups/search:
    get:
      description: Search groups with keyword and pagination
      parameters:
      - description: Search keyword
        in: query
        name: keyword
        required: true
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/middleware.APIResponse'
            - properties:
                data:
                  additionalProperties:
                    allOf:
                    - {}
                    - properties:
                        groups:
                          items:
                            $ref: '#/definitions/controller.GroupResponse'
                          type: array
                        total:
                          type: integer
                      type: object
                  type: object
              type: object
      security:
      - ApiKeyAuth: []
      summary: Search groups
      tags:
      - groups
...
swagger: "2.0"